+++ apps/meteor/server/services/video-conference/service.ts
@@ -501,13 +501,18 @@ export class VideoConfService extends ServiceClassInternal implements IVideoConf
 			msg: '',
 			groupable: false,
 			blocks: customBlocks || [this.buildVideoConfBlock(call._id)],
-		};
+		} satisfies Partial<IMessage>;
 
 		const room = await Rooms.findOneById(call.rid);
 		const appId = videoConfProviders.getProviderAppId(call.providerName);
 		const user = createdBy || (appId && (await Users.findOneByAppId(appId))) || (await Users.findOneById('rocket.cat'));
 
 		const message = await sendMessage(user, record, room, false);
+
+		if (!message) {
+			throw new Error('failed-to-create-message');
+		}
+
 		return message._id;
 	}
 
@@ -613,7 +618,7 @@ export class VideoConfService extends ServiceClassInternal implements IVideoConf
 		await Push.send({
 			from: 'push',
 			badge: 0,
-			sound: 'default',
+			sound: 'ringtone.mp3',
 			priority: 10,
 			title: `@${call.createdBy.username}`,
 			text: i18n.t('Video_Conference'),