+++ packages/fuselage-ui-kit/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 3.0.3
+
+### Patch Changes
+
+- @rocket.chat/gazzodown@3.0.3
+- @rocket.chat/ui-contexts@3.0.3
+- @rocket.chat/ui-video-conf@3.0.3
+
+## 3.0.2
+
+### Patch Changes
+
+- @rocket.chat/gazzodown@3.0.2
+- @rocket.chat/ui-contexts@3.0.2
+- @rocket.chat/ui-video-conf@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.1
+- @rocket.chat/gazzodown@3.0.1
+- @rocket.chat/ui-video-conf@3.0.1
+
 ## 3.0.0
 
 ### Patch Changes